Offer more clarity on incident response (and some minor wordsmithing) (#57)
We've received feedback that it'd be useful, especially for potential new log operators, to have more clarity on log operator expectations during incidents, and what Chrome's responses might be. This change tries to offer that.

Most of this should be relatively non-controversial, but it does include a explicit few changes to policy: that we consider 90-day availability dropping below 95%, or persistent log availability issues lasting more than two weeks, to be sufficiently severe to justify log removal.

It also shifts the policy towards not capitalizing actors in the ecosystem (e.g. "log operator" vs "Log Operator") and some other smaller wordsmithing.